Move settings bundles into own file
